Add Adult to LearningTest.
Used in KNITest
