+++ .github/actions/test-kni/action.yml
@@ -6,14 +6,12 @@ runs:
   steps:
     - name: Check ldconfig
       shell: bash
+      if: runner.os == "Linux"
       run: |
-        if ["$RUNNER_OS" == "Linux"]
+        if [ $(ldconfig -p | grep -c KhiopsNativeInterface) -eq 0 ] 
         then
-          if [ $(ldconfig -p | grep -c KhiopsNativeInterface) -eq 0 ] 
-          then
-            echo "::error::KNI is not reachable with ld, check post installation"
-            false
-          fi
+          echo "::error::KNI is not reachable with ld, check post installation"
+          false
         fi
     - name: Checkout sources
       uses: actions/checkout@v3
@@ -24,36 +22,35 @@ runs:
       with:
         repository: KhiopsML/KhiopsNativeInterface-tutorial
         path: tutorial
-    - name: Compile C examples
+    - name: Compile C examples (Linux)
       shell: bash
+      if: runner.os == "Linux"
       run: |
-        if [ "$RUNNER_OS" == "Linux" ] ;
-        then
-          gcc -o KNIRecodeFile tutorial/cpp/KNIRecodeFile.c -I khiops/src/Learning/KhiopsNativeInterface/ \
+        gcc -o KNIRecodeFile tutorial/cpp/KNIRecodeFile.c -I khiops/src/Learning/KhiopsNativeInterface/ \
                 -lKhiopsNativeInterface  -ldl
-          gcc -o KNIRecodeMTFiles tutorial/cpp/KNIRecodeMTFiles.c  -I khiops/src/Learning/KhiopsNativeInterface/ \
+        gcc -o KNIRecodeMTFiles tutorial/cpp/KNIRecodeMTFiles.c  -I khiops/src/Learning/KhiopsNativeInterface/ \
                 -lKhiopsNativeInterface -ldl
-        fi
-        if [ "$RUNNER_OS" == "Windows" ] ;
-        then
-          call "C:\\Program Files\\Microsoft Visual Studio\\2022\\Enterprise\\VC\\Auxiliary\\Build\\vcvars64.bat"
-          cl tutorial/cpp/KNIRecodeFile.c -I khiops/src/Learning/KhiopsNativeInterface/ KhiopsNativeInterface64.lib /link "/LIBPATH:kni"
-          cl tutorial/cpp/KNIRecodeMTFiles.c  -I khiops/src/Learning/KhiopsNativeInterface/ KhiopsNativeInterface64.lib /link "/LIBPATH:kni"
-        fi
+    - name: Compile C examples (Windows)
+      if: runner.os == "Windows"
+      shell: cmd
+      run: |
+        call "C:\\Program Files\\Microsoft Visual Studio\\2022\\Enterprise\\VC\\Auxiliary\\Build\\vcvars64.bat"
+        cl tutorial/cpp/KNIRecodeFile.c -I khiops/src/Learning/KhiopsNativeInterface/ KhiopsNativeInterface64.lib /link "/LIBPATH:kni"
+        cl tutorial/cpp/KNIRecodeMTFiles.c  -I khiops/src/Learning/KhiopsNativeInterface/ KhiopsNativeInterface64.lib /link "/LIBPATH:kni"
     - name: Test mono-table example
       shell: bash
-      run: |-
+      run: |
         ./KNIRecodeFile tutorial/data/ModelingIris.kdic SNB_Iris tutorial/data/Iris.txt R_Iris.txt
         if [ ! -f R_Iris.txt ] ;
         then
           echo "::error::R_Iris.txt is missing"
           false
         fi
-        if diff -q R_Iris.txt khiops/packaging/common/KNI/R_Iris.txt > /dev/null ; 
+        if diff -q R_Iris.txt khiops/packaging/common/KNI/results.ref/R_Iris.txt > /dev/null ; 
         then
-          echo "Mono table recodage is OK"
+          echo "Mono table recoding is OK"
         else
-        echo "::error::Recodage issue with Iris.txt"
+        echo "::error::Recoding issue with Iris.txt"
           false
         fi
     - name: Test multi-tables example
@@ -66,10 +63,10 @@ runs:
           echo "::error::R_SpliceJunction.txt is missing"
           false
         fi
-        if diff -q R_SpliceJunction.txt khiops/packaging/common/KNI/R_SpliceJunction.txt > /dev/null ; 
+        if diff -q R_SpliceJunction.txt khiops/packaging/common/KNI/results.ref/R_SpliceJunction.txt > /dev/null ; 
         then
-          echo "Multi table recodage is OK"
+          echo "Multi table recoding is OK"
         else
-          echo "::error::Recodage issue with SpliceJunction.txt"
+          echo "::error::Recoding issue with SpliceJunction.txt"
           false
         fi
