+++ .github/workflows/update-kni-tutorial.yml
@@ -22,8 +22,8 @@ jobs:
           repository: KhiopsML/KhiopsNativeInterface-tutorial
           ssh-key: ${{ secrets.KNI_TUTORIAL_SSH_KEY }}
           path: tutorial
-    # Copy sources files and README.md from khiops repository to KhiopsNativeInterface-tutorial
-    # README.md and *.c files are generated by cmake configure
+      # Copy sources files and README.md from khiops repository to KhiopsNativeInterface-tutorial
+      # README.md and *.c files are generated by cmake configure
       - name: Update local KhiopsNativeInterface-tutorial
         run: |
           cmake -B khiops/build -S khiops -DMPI=OFF -DBUILD_JARS=OFF -DBUILD_LEX_YACC=OFF -DTESTING=OFF 
@@ -33,6 +33,7 @@ jobs:
           cp khiops/src/Learning/KhiopsNativeInterface/KhiopsNativeInterface.h tutorial/include/
           cp khiops/src/Learning/KNITransfer/KNIRecodeFile.h tutorial/cpp/
           cp khiops/src/Learning/KNITransfer/KNIRecodeMTFiles.h tutorial/cpp/
+          cp packaging/common/KNI/java/* tutorial/java/
       - name: Update remote KhiopsNativeInterface-tutorial
         run: |-
           cd tutorial
