+++ test/KNITest/results.ref/KNI_Iris.txt
@@ -1,0 +1,34 @@
+Begin test KNI
+SYS KNI version: 110
+SYS KNI full version: 10.1.5
+Open stream @ROOT_DIR@/test/KNITest/ModelingIris.kdic SNB_Iris: 1
+Recode stream record Iris-setosa	Iris-setosa	0.9960418693	0.001979090503	0.001979040221: 0
+Recode stream record Iris-virginica	Iris-virginica	0.00198971666	0.03512915094	0.9628811324: 0
+Recode invalid stream record : -14
+Recode empty stream record : -14
+Recode NULL stream record : -13
+Recode stream record without enough output length : -15
+Recode stream record with NULL output record : -15
+Recode stream record with negative output size : -15
+
+Wrong parameter tests
+Open stream with NULL dictionary file NULL SNB_Iris: -2
+Open stream with missing dictionary file WrongFile SNB_Iris: -3
+Open stream with invalid dictionary file @ROOT_DIR@/test/KNITest/../LearningTest/datasets/Iris/Iris.txt SNB_Iris: -4
+Open stream with NULL dictionary @ROOT_DIR@/test/KNITest/ModelingIris.kdic NULL: -5
+Open stream with missing dictionary @ROOT_DIR@/test/KNITest/ModelingIris.kdic WrongDic: -6
+Open stream with NULL header line @ROOT_DIR@/test/KNITest/ModelingIris.kdic SNB_Iris: -8
+Open stream with bad header line @ROOT_DIR@/test/KNITest/ModelingIris.kdic SNB_Iris: -23
+Open stream with eol separator @ROOT_DIR@/test/KNITest/ModelingIris.kdic SNB_Iris: -9
+Recode with wrong stream : -10
+Close wrong stream (-1): -10
+Close wrong stream (0): -10
+
+Muliple test: up to 1000 simultaneous streams
+SYS Muliple open stream	317: Error -22
+SYS Muliple close stream	317: Error -10
+
+Recode records of @ROOT_DIR@/test/KNITest/../LearningTest/datasets/Iris/Iris.txt to @ROOT_DIR@/test/KNITest/results/R_Iris.txt
+Recoded record number: 150
+
+End test KNI
