+++ test/LearningTest/datasets/Adult/Adult.kdic
@@ -1,0 +1,21 @@
+#Khiops 7.7.2i
+
+Dictionary	Adult
+{
+	Numerical	Label		;	
+	Numerical	age		;	
+	Categorical	workclass		;	
+	Numerical	fnlwgt		;	
+	Categorical	education		;	
+	Numerical	education_num		;	
+	Categorical	marital_status		;	
+	Categorical	occupation		;	
+	Categorical	relationship		;	
+	Categorical	race		;	
+	Categorical	sex		;	
+	Numerical	capital_gain		;	
+	Numerical	capital_loss		;	
+	Numerical	hours_per_week		;	
+	Categorical	native_country		;	
+	Categorical	class		;	
+};
